pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "telescope"
version = "1.0.1"
description = "A tool to observe distant (or local!) Airflow installations, and gather metadata or other required data."
readme = "README.md"
authors = ["telescope <solutions@astronomer.io>"]
repository = "https://github.com/telescope/telescope"
homepage = "https://github.com/telescope/telescope"
include = ["telescope/config.yaml"]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"telescope" = "telescope.__main__:cli"

[tool.poetry.dependencies]
python = "^3.8"
importlib_metadata = {version = "^4.5.0", python = "<3.8"}
rich = "^10.7.0"
kubernetes = "^19.15.0"
fabric = "^2.6.0"
docker = "^5.0.3"
invoke = "^1.6.0"
PyYAML = "^6.0"
click = "^8.0.3"
tqdm = "^4.62.3"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = {version = "^21.7b0", allow-prereleases = true}
darglint = "^1.8.0"
isort = {extras = ["colors"], version = "^5.9.3"}
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pre-commit = "^2.14.0"
pydocstyle = "^6.1.1"
pylint = "^2.10.2"
pytest = "^6.2.4"
pyupgrade = "^2.24.0"
safety = "^1.10.3"
pytest-xdist = "^2.4.0"
pytest-integration = "^0.2.2"
pytest-mock = "^3.6.1"
apache-airflow = "^2.2.1"

[tool.black]
# https://github.com/psf/black
target-version = ["py38"]
line-length = 120
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | tests/
    | telescope/airflow_report.py
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 38
line_length = 120

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.8
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--ignore=telescope/airflow_report.py",
  "--disable-warnings",
  "--no-header",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
